requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.38.36
botocore==1.38.36
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
Deprecated==1.2.18
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.2.3
h11==0.16.0
idna==3.10
jmespath==1.0.1
jwt==1.3.1
lxml==5.4.0
oracledb==3.1.1
packaging==25.0
pikepdf==9.8.1
pillow==11.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.6
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
wrapt==1.17.2
pydantic[email]
uvicorn[standard]
gunicorn==23.0.0
orjson==3.10.18